requirements.txt
streamlit
transformers
requests
langchain
langchain_groq
python-dotenv
